Extracted TestSettings utility
