+++ Plan/nukkit/src/test/java/com/djrapitops/plan/NukkitSystemTest.java
@@ -17,7 +17,6 @@
 package com.djrapitops.plan;
 
 import com.djrapitops.plan.exceptions.EnableException;
-import com.djrapitops.plan.settings.ConfigSettingKeyTest;
 import com.djrapitops.plan.settings.config.PlanConfig;
 import com.djrapitops.plan.settings.config.paths.WebserverSettings;
 import com.djrapitops.plan.settings.config.paths.key.Setting;
@@ -25,6 +24,7 @@ import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.io.TempDir;
 import utilities.RandomData;
+import utilities.TestSettings;
 import utilities.mocks.NukkitMockComponent;
 
 import java.nio.file.Path;
@@ -65,8 +65,8 @@ class NukkitSystemTest {
             system.enable();
             PlanConfig config = system.getConfigSystem().getConfig();
 
-            Collection<Setting> serverSettings = ConfigSettingKeyTest.getServerSettings();
-            ConfigSettingKeyTest.assertValidDefaultValuesForAllSettings(config, serverSettings);
+            Collection<Setting> serverSettings = TestSettings.getServerSettings();
+            TestSettings.assertValidDefaultValuesForAllSettings(config, serverSettings);
         } finally {
             system.disable();
         }
